Update en_us.json
Found a couple grammatical errors that ought to be corrected.

Use "an" instead of "a" when the following word begins with a vowel sound. "An Afrit" instead of "A Afrit"

Fixed some missing capitalization. Generally in the book spirit types are written as proper nouns (Foliot instead of foliot).

For Lime Chalk the title felt out of place compared to the others, so it was made "refreshing" as lime drinks are refreshing.

Orange Chalk title changed. At the very least this would be "Hot Little" not "Little Hot". 

Green Chalk title similarly felt out of place, like it couldn't decide. Settled to "Chalk of the Wilds".

Light Blue chalk title changed to Glacial Chalk. Glaced is not a word in English. 

Following the "Spirit Summoning" format for summoning rituals, the others are changed to "Spirit Possession" and "Spirit Infusion" for consistency and clarity. In the case of possession, the spirit is possessing the object or creature. As written, it meant the spirit was being possessed by the object or creature. 

There are a few more in there for clarity's sake. Feel free to ask if you have any questions on why they were changed.